Move stripe tests to a "routers" package
